Redo the way certain large list variables are populated to allow pre-processing
The goal here is to switch from "one large YAML statement with many ifs"
to "separate if statements".

The first doesn't lend itself to having pre-processing comments around
it and to actually pre-processing it based on these comments.

Pre-processing is a future goal I'm working toward.
The idea is that `if` statements for component X do not need to run
if you're not using component X. In the future, we should be able to
have the playbook auto-configure itself based on the components you're using
and skip a lot of unnecessary work.
This is one step toward achieving this goal.