Update authentik.md: instruction for configuring a shared Valkey instance
Copied from https://github.com/mother-of-all-self-hosting/mash-playbook/blob/00040cd450f5c8abea36d940410be52729b560ec/docs/services/funkwhale.md and edit it.

This change also reflects the recommendation by the playbook to set up mash.example.com as the main hostname for its A record and set subdomains to CNAME records. Indeed it is not quite realistic to presume that a user would set authentik.example.com as the sole hostname which is to be set to the A record against the recommendation.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>