Add versioning=semver to renovate.json
By default Renovate adopts PEP440 versioning for Galaxy Collection as the datasource. According to the Python Packaging User Guide, the pre-release segment MUST be "{a|b|rc}N", therefore updates like from x.x.x-0 to x.x.x-1 are not detected even if `ignoreUnstable` is set to false on renovate.json. Here Semver is specified to versioning, as it allows to denote a pre-release version by appending a hyphen.

Resources:
- https://docs.renovatebot.com/modules/datasource/galaxy-collection/
- https://packaging.python.org/en/latest/specifications/version-specifiers/#public-version-identifiers
- https://semver.org/spec/v2.0.0.html#spec-item-9

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>