Update renovate.json: replace versioning=semver with versioning=loose
On renovate.json are there roles which do not follow SemVer strictly, and it has stopped the likes of `clickhouse` and `radicale` from being correctly detected. The documentation at Renovate claims that `loose` versioning could be used as a fallback for SemVer. With `loose`, it also looks like roles which do not have .PATCH version can be detected as well, such as `healthchecks` (v3.10-2) and `jitsi` (v10314-1).

References:
- https://docs.renovatebot.com/modules/versioning/loose/
- https://docs.renovatebot.com/configuration-options/#versioning

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>