Only add Traefik labels for exposing metrics for APISIX Gateway & Promtail when `mash_playbook_metrics_exposure_enabled`, not when `prometheus_enabled`
Adding Traefik labels is for public metrics exposure.

The fact that Prometheus is enabled (`prometheus_enabled`) is a sign
that the user likely wants metrics enabled and exposed on the container
network, so that this Prometheus instance can reach them.
However, it's not an indication that metrics are to be exposed publicly.

Only `mash_playbook_metrics_exposure_enabled` is an indication of this.

Fixes https://github.com/mother-of-all-self-hosting/mash-playbook/issues/857