Allow encoded characters in traefik for collabora online (#1543)
* Allow encoded characters in traefik for collabora online

Traefik v3.6.4 introduced a more restrictive way of handling encoded characters in paths:
https://doc.traefik.io/traefik/migrate/v3/#v364
Which lead to an inaccessible Collabora Online websocket. Adding allowEncodedSlash and allowEndodedQuestionMark options in traefik fixes this issue.

* Fix typo

* Fix another typo

* Make `encodedCharacters` Traefik overrides more extensible

A `role-specific` block is OK if only Collabora Online will ever need this.

There will likely be other servces that will need a an `encodedCharacters` tweak here or there, so we'd better not use a `role-specific` block.

---------

Co-authored-by: Slavi Pantaleev <slavi@devture.com>